app/requirements.txt
annotated-types==0.7.0
APIFlask==2.1.1
apispec==6.6.1
asttokens==2.4.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
executing==2.0.1
Flask==3.0.3
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==2.6.0
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-oneofschema==3.1.1
marshmallow_dataclass==8.6.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
neo4j==5.20.0
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
psycopg2-binary==2.9.9
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
python-arango==8.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==70.0.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webargs==8.4.0
Werkzeug==3.0.3
zipp==3.19.1
